clean up integration tests
